Restructure models so that xgboost and Keras are independent
