+++ gentun/__init__.py
@@ -1,12 +1,28 @@
+import sys
+import os
+
+# Avoid printing 'Using TensorFlow
+stderr = sys.stderr
+sys.stderr = open(os.devnull, 'w')
+
 # Make public APIs available at top-level import
 from .algorithms import GeneticAlgorithm
 from .populations import Population, GridPopulation
 from .master import DistributedPopulation, DistributedGridPopulation
 from .worker import GentunWorker
 
-# Individual-Model pairs
+# xgboost individuals and models
 try:
     from .individuals import XgboostIndividual
-    from .models import XgboostModel
+    from .models.xgboost_models import XgboostModel
 except ImportError:
     print("Warning: install xgboost to use XgboostIndividual and XgboostModel.")
+
+# Keras individuals and models
+try:
+    from .individuals import GeneticCnnIndividual
+    from .models.keras_models import GeneticCnnModel
+except ImportError:
+    print("Warning: install Keras and TensorFlow to use GeneticCnnIndividual and GeneticCnnModel.")
+
+sys.stderr = stderr
